+++ src/Smartstore.Web/Models/ShoppingCart/CartModelBase.cs
@@ -40,6 +40,7 @@ namespace Smartstore.Web.Models.Cart
 
         public int EnteredQuantity { get; set; }
         public LocalizedValue<string> QuantityUnitName { get; set; }
+        public LocalizedValue<string> QuantityUnitNamePlural { get; set; }
         public List<SelectListItem> AllowedQuantities { get; set; } = new();
         public int MinOrderAmount { get; set; }
         public int MaxOrderAmount { get; set; }